+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+* Added `collections` attribute to list of default fields to include, so that we satisfy the STAC API spec, which requires a `collections` attribute to be output when an item is part of a collection
 ### Removed
 
 ### Changed